+++ app/filesystem/configurationWriter.js
@@ -4,16 +4,30 @@ function configurationWriter(
     process
 ) {
     'use strict';
-    
-    function writeConfigFile(fileName, fileContent) {
+
+    function buildDestinationPath(fullFileName) {
         const destinationDirectory = process.cwd();
-        const destinationPath = path.join(destinationDirectory, `${fileName}.js`);
+
+        return path.join(destinationDirectory, fullFileName);
+    }
+
+    function writeContentToConfig(fullFileName, fileContent) {
+        const destinationPath = buildDestinationPath(fullFileName);
 
         fs.writeFileSync(destinationPath, fileContent, { encoding: 'utf8' });
     }
 
+    function writeConfigFile(fileName, fileContent) {
+        writeContentToConfig(`${fileName}.js`, fileContent);
+    }
+
+    function writeJSONConfigFile(fileName, fileContent) {
+        writeContentToConfig(`${fileName}.json`, fileContent);
+    }
+
     return {
-        writeConfigFile: writeConfigFile
+        writeConfigFile: writeConfigFile,
+        writeJSONConfigFile: writeJSONConfigFile
     };
 }