+++ readme/content/version-history.md
@@ -3,6 +3,11 @@
     (title "Version History"))
 /bl-->
 
+**v1.1.0**
+
+- Updated default file location to app instead of dependencies for configuration builders
+- Introduced configuration construction for ES-Next Module DI container builder -- `dject --configure-es-module-builder`
+
 **v1.0.0**
 
 Initial release. Current feature support: